Improve `Manager` usability and remove `Launcher` (#131)
* Merge launcher and manager interfaces

* Update examples to use new manager

* Fix race with cancelled futures and closing clients

One of the handle tests exposed a race where an action is requested via
a handle, the handle is closed with cancel_futures=True, and then later
the exchange client receives the result for the action and tries to set
the result on the future created by the handle, but this raises and
invalid state error because it was cancelled.

* Lift methods into manager for ease-of-use

* Improve agent exception raising on manager close

* Update README and docs

* Support waiting on multiple agents in manager

* Restore runtime_checkable to Handle

+++ testing/redis.py
@@ -18,7 +18,7 @@ class MockRedis:
     async def blpop(
         self,
         keys: list[str],
-        timeout: int = 0,
+        timeout: float = 0,
     ) -> list[str] | None:
         result: list[str] = []
         for key in keys:
